Lookahead size must now be < window size, not <= .
Using a lookahead size equal to the window size could previously lead to
an infinite loop (discovered and fixed by @unixdj, thanks). There really
isn't any benefit to having a lookahead size that large, as it only
applies when the entire input stream is the same. In all other cases, it
makes the compression ratio worse.

Closes #20.