Add option to override heatshrink configuration
Currently its impossible to change heatshrink compression configuration
wihout touching the library header files. This patch makes it possible
for configuration which can be configured at compilation time without
having to touch the library. Which makes it easier to keep it sync with
github source without modifying the actual source.

Signed-off-by: Ajay Bhargav <contact@rickeyworld.info>